Plot variance explained by PCA components

Plot the variance explained by each PCA component. The number of components shown is controlled by num_components, with up to the first 10 componenets shown by default. If less than the requested number of components exist, then all the components will be shown.

plot_pca_variance(pca, num_components = 10)

Arguments

pca

output of pca_wsaf() function.

num_components

maximum components to be shown.